correctly synchronize stores/loads in futer
To correctly synchronize atomic operations on the lock, a more strict
ordering than Relaxed is needed.

Relaxed allows for reording of stores/loads to other memory regions,
before or after the atomic operation. Thus, for example, stores in the
critical section, could be reordered before acquiring the lock,
completely bypassing any protections that the futex provides.

To fix that, we use Release ordering on stores, and Acquire ordering on
loads.