Use compare_exchange_weak in loops
Some platforms do no provide one single instruction for
compare_exchange. Thus, the compare_exchange operation is implemented by
using other instructions in a loop, until the operation as a whole
succeeds.

Since we use compare_exchange in a loop, this would cause a nested loop
structure, which in some cases is not efficient. Instead, use
compare_exchange_weak which can spuriously fail, but does not contain a
loop.