use dependency inversion in Futer
To be able to mock the futex implementation used in the Futer type, we
need to use the dependency inversion pattern.

To do that, we create internal types that are also generic over possible
futex implementations, and use methods provided by a trait interface.
Then a concrete implementation of the futex simply directly calls
futex_wake and futex_wait.